+++ packages/react/src/components/dropdownComponents/modularOptionList/components/SingleSelectAndGrouplessList.tsx
@@ -59,9 +59,11 @@ export function SingleSelectAndGrouplessList() {
   const { getData, trigger, getMetaData } = useModularOptionListDataHandlers();
   const { groups, multiSelect } = getData();
   const { getOptionId, refs, elementIds } = getMetaData();
-  const attr = createListElementProps({ refs, elementIds, multiSelect });
-  const shouldRenderOptions = true; // TODO !isSearching;
+  const attr = {
+    ...createListElementProps({ refs, elementIds, multiSelect }),
+    'aria-live': 'polite' as const,
+  };
 
-  const children = shouldRenderOptions ? createOptionElements({ groups, trigger, multiSelect, getOptionId }) : null;
+  const children = createOptionElements({ groups, trigger, multiSelect, getOptionId });
   return <ul {...attr}>{children}</ul>;
 }