+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - [Component] What is added?
+- [Select] Add 'aria-live'-attribute to search/filter results on the Select component.
 
 #### Changed
 
@@ -129,6 +130,7 @@ Changes that are not related to specific components
 #### Fixed
 
 - [Component] What bugs/typos are fixed?
+
 ## [4.4.0] - April, 3rd, 2025
 
 ### React